lessons/01-text-chatbot/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
openai==2.21.0
pydantic==2.10.6
python-dotenv==1.0.1
